+++ rs/canister/api/src/queries/get_principal.rs
@@ -1,0 +1,7 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct GetPrincipalArgs {
+    pub email: String,
+}